# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Luna DevOps API
  version: 0.1.0
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/oauth/device/authorization:
    post:
      operationId: startOAuthDeviceAuthorization
      tags: [Auth]
      summary: Start an OAuth Device Authorization Grant
      x-luna-cli:
        classification: protocol-adapter
        hidden: true
        exclusionReason: Device authorization is exposed through `luna login`.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [client_id]
              properties:
                client_id: { type: string }
                scope: { type: string }
      responses:
        "200":
          description: Device and user codes for the browser verification flow.
          content:
            application/json:
              schema:
                type: object
                required: [device_code, user_code, verification_uri, verification_uri_complete, expires_in, interval]
                properties:
                  device_code: { type: string }
                  user_code: { type: string }
                  verification_uri: { type: string, format: uri }
                  verification_uri_complete: { type: string, format: uri }
                  expires_in: { type: integer, minimum: 1 }
                  interval: { type: integer, minimum: 1 }
        "400":
          description: OAuth protocol error.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OAuthProtocolError" }
  /api/v1/oauth/device/verification:
    get:
      operationId: getOAuthDeviceVerification
      tags: [Auth]
      summary: Inspect a pending OAuth device authorization
      x-luna-cli:
        classification: protocol-adapter
        hidden: true
        exclusionReason: Browser-only Device Code verification endpoint.
      parameters:
        - name: user_code
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Pending device authorization visible to the signed-in user.
          content:
            application/json:
              schema:
                type: object
                required: [application, userCode, scopes, expiresAt]
                properties:
                  application: { $ref: "#/components/schemas/OAuthApplication" }
                  userCode: { type: string }
                  scopes:
                    type: array
                    items: { type: string }
                  expiresAt: { type: string, format: date-time }
        "400":
          description: Invalid, expired, or consumed user code.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    post:
      operationId: decideOAuthDeviceVerification
      tags: [Auth]
      summary: Approve or deny an OAuth device authorization
      x-luna-cli:
        classification: protocol-adapter
        hidden: true
        exclusionReason: Browser-only Device Code verification endpoint.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userCode, decision]
              properties:
                userCode: { type: string }
                decision: { type: string, enum: [approve, deny] }
      responses:
        "200":
          description: Device authorization decision accepted.
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status: { type: string, enum: [approved, denied] }
        "400":
          description: Invalid, expired, or consumed user code.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: lyd_session
      description: HttpOnly browser session cookie. Sensitive interactive operations do not accept personal access tokens as a substitute.
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: Luna OAuth access token or personal access token
      description: Luna OAuth access token or personal access token. Individual operations may require OAuth and explicitly reject personal access tokens.
    AIAgentServiceToken:
      type: http
      scheme: bearer
      bearerFormat: independent Agent callback service credential
    AIDelegationToken:
      type: http
      scheme: bearer
      bearerFormat: 60-second operation-bound AI delegation
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: Stable machine-readable error code.
        error:
          type: string
          description: Safe user-facing error message.
        detail:
          type: string
          description: Development-only diagnostic detail.
        purpose:
          type: string
          description: Step-up purpose returned with `mfa_required`.
        details:
          type: object
          additionalProperties: true
          description: Optional machine-readable error context. Scope failures include `requiredScope`.
      required: [code, error]
    OAuthApplication:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        clientId: { type: string }
        redirectUris:
          type: array
          items: { type: string, format: uri }
        scopes:
          type: array
          items: { type: string }
        confidential: { type: boolean }
        active: { type: boolean }
      required: [id, name, clientId, redirectUris, scopes, confidential, active]
...
    OAuthProtocolError:
      type: object
      properties:
        error:
          type: string
          enum:
            - invalid_request
            - invalid_client
            - invalid_grant
            - unauthorized_client
            - unsupported_grant_type
            - invalid_scope
            - authorization_pending
            - slow_down
            - access_denied
            - expired_token
        error_description: { type: string }
      required: [error]
...